po: simplify xgettext rule
Every time xgettext is run, the generated template has its own
POT-Creation-Date field with the runtime timestamp; because of this,
the comparison between the existing template and the new one (using
`cmp`) always results as "different".

Hence, simply run xgettext with the wanted output file, as the result
is the same; make use of the automatic make variables to avoid repeating
variables in the command invocation.