build: stop updating the catalog on 'all'
While it can be useful to keep the translation catalog up-to-date,
updating it every time a file is changed is a bit wasteful.

Hence, do not ensure that the pot file is updated on 'all'. This will
likely require some manual update from time to time to make sure that
the translatable messages are still up-to-date, however they do not
change that often in linuxlogo.